requirements/auth.txt
Flask==1.1.4
Flask-SQLAlchemy
gunicorn
flask-restx
pydantic
black
isort
redis
pyjwt
psycopg2-binary
passlib
user-agents
requests
Authlib
pydantic[dotenv]
multicolorcaptcha
sqlalchemy-stubs
python-logstash
sentry-sdk[flask]